requirements.txt
conda==4.10.3
conda-package-handling @ file:///Users/runner/miniforge3/conda-bld/conda-package-handling_1618231472205/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851523607/work
cycler==0.11.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1627075104865/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
distlib==0.3.2
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
Fiona==1.8.21
fonttools==4.32.0
gcloud==0.18.3
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
googleapis-common-protos==1.56.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
jmespath==0.10.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1624941339085/work
munch==2.5.0
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1628690102344/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1629092076711/work
oauth2client==4.1.3
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
outcome==1.1.0
pandas==1.4.1
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1625678080886/work
platformdirs==2.3.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1629903925368/work
protobuf==3.20.1rc1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyproj==3.3.0
pyspark==3.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-decouple==3.6
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
PyYAML==5.4.1
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1628233184753/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel-yaml-conda @ file:///Users/runner/miniforge3/conda-bld/ruamel_yaml_1624999685134/work
s3transfer==0.5.0
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
Shapely==1.8.1.post1
sortedcontainers==2.4.0
sqlparse==0.4.2
torch==1.9.0
torchvision==0.2.2
trio-websocket==0.9.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
virtualenv==20.7.2
webencodings==0.5.1
websocket-client @ file:///Users/runner/miniforge3/conda-bld/websocket-client_1610127894508/work
